Make sure we don't rebuild the docs and QA datasets
Indeed, when cloning the repository for the first time,
the timestamps were not in the right order between files,
causing the make command to redownload and scrap the documents dataset
and rebuild the QA dataset.

We now check whether they are already on disk,
and skip doing those steps if they are.